pyproject.toml
[tool.poetry]
name = "blacklist_tg_bot"
version = "0.1.0"
description = ""
authors = ["MargoRSq <49024821+MargoRSq@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.7"
starlette = "^0.15.0"
autopep8 = "^1.5.7"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
